requirements.txt
flask==3.1.0
pymongo==4.10.1
werkzeug==3.1.3
google-generativeai==0.8.3
groq==0.13.0
opencv-python==4.10.0.84
mediapipe
numpy==2.1.3
python-dotenv==1.0.1
dnspython==2.7.0
